Refactor: split fields.tsx, terminal-pane.tsx, main.tsx into focused modules (F8)
Behaviour-preserving structural split of the three largest renderer files, per the documented -parts/ convention. Code moved verbatim; zero logic/reactivity/styling change.

fields.tsx (1008 LOC) -> an 8-line barrel re-exporting the same public set, with the 12 components grouped into settings-modal-parts/fields/{primitives,theme,card-density,action-template,color,shortcut,terminal}.tsx (all import sites unchanged).

terminal-pane.tsx (905 -> 257) -> the god controller (session reconciliation + the 8-method imperative handle + xterm/IPC/resize/drag wiring) extracted verbatim into terminal-pane/controller.ts's createTerminalController hook, called synchronously from the component body so Solid signal/effect ownership and the imperative surface are identical.

main.tsx (972 -> 831) -> the ~194-line modal/overlay tail (13 modal <Show> blocks + preview + toast) extracted into modal-host.tsx, props typed by indexed access from the owning hooks so the pass-through is exact.

Gate: typecheck clean, 1286 unit, build green, and Playwright e2e 48 passing (incl. terminal-open, settings-modal-audit, modal-back-stack, screenshots light+dark).
